Fixed Lombok problems builder(). Edited Lombok dependencies in the pom.xml Changed type of bid to int Wrote code in addNewCampaign method Added annotations @Document(repository) in the Campaign

+++ src/main/java/com/mabaya/exam/app/controller/AppController.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDate;
+
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 @RestController
 @RequestMapping//("/mabaya")
@@ -21,9 +22,10 @@ public class AppController {
     ProductService ProductService;
 
     @PostMapping("/createCampaign/{name}/{startDate}/{category}/{bid}")
-    public CampaignDto createCampaign(@PathVariable String name, @PathVariable LocalDate startDate,
-                                      @PathVariable String category, @PathVariable Double bid) {
-        return CampaignService.addNewCampaign(name,startDate,category,bid);
+    public CampaignDto createCampaign(
+            @PathVariable String name, @PathVariable LocalDate startDate,
+            @PathVariable String category, @PathVariable int bid) {
+        return CampaignService.addNewCampaign(name, startDate, category, bid);
     }
 
     @GetMapping("/serveAd/{category}")
